Update version to 0.268.0 and refine cost name handling in NEAT framework
- Bumped version in deno.json to 0.268.0.
- Simplified type exports in mod.ts by removing unnecessary types from NeatOptions.
- Introduced a new CostName type in Costs.ts to restrict cost names to built-in values, enhancing type safety.
- Updated TrainOptions and DiscoveryRunner to utilize the new CostName type for costName properties.
- Adjusted tests to ensure only built-in cost names are accepted, improving validation.

These changes enhance the clarity and robustness of cost name handling in the NEAT framework.
