+++ mod.ts
@@ -52,10 +52,7 @@ export { CreatureUtil } from "./src/architecture/CreatureUtils.ts";
  *
  * @see {@link module:src/config/NeatOptions}
  */
-export type {
-  DiscoveryMinCandidatesPerCategory,
-  NeatOptions,
-} from "./src/config/NeatOptions.ts";
+export type { NeatOptions } from "./src/config/NeatOptions.ts";
 
 /**
  * Cost Interface
